refactor: change names styled-components and path (Fri, October 21, 2022)
