+++ components/src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Navigate, Route, Routes } from 'react-router-d
 import Layout from './components/Layout/Layout';
 import Main from './pages/Main/Main';
 import About from './pages/About/About';
-import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
+import ErrorPage from './pages/ErrorPage/ErrorPage';
 import Forms from './pages/Forms/Forms';
 
 function App() {
@@ -15,7 +15,7 @@ function App() {
             <Route index element={<Main />} />
             <Route path="about" element={<About />} />
             <Route path="forms" element={<Forms />} />
-            <Route path="404" element={<NotFoundPage />} />
+            <Route path="404" element={<ErrorPage />} />
             <Route path="*" element={<Navigate to="404" replace />} />
           </Route>
         </Routes>
