+++ components/src/components/MainCard/MainCard.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { CardStyled } from '../../pages/Main/styled';
 import { CardProps } from '../../pages/Main/Main.model';
+import { MainCardStyled } from './styled';
 
-const Card = (props: CardProps) => {
+const MainCard = (props: CardProps) => {
   const { Poster, Title, Type, Year, imdbID } = props.movie;
 
   return (
-    <CardStyled data-testid="Card" onClick={() => props.setModalId(imdbID)}>
+    <MainCardStyled data-testid="MainCard" onClick={() => props.setModalId(imdbID)}>
       <figure>
         <img src={Poster} alt={Title} />
         <h3>{Year}</h3>
@@ -15,8 +15,8 @@ const Card = (props: CardProps) => {
       <figcaption>
         <h2>{Title}</h2>
       </figcaption>
-    </CardStyled>
+    </MainCardStyled>
   );
 };
 
-export default Card;
+export default MainCard;
