+++ components/src/components/ValidErorr/styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ValidErrorStyled = styled.p`
+  font-size: 0.7em;
+  color: ${({ theme }) => theme.colors.red};
+  text-align: center;
+`;
