+++ components/src/pages/ErrorPage/ErrorPage.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { ErrorMainStyle } from './styled';
 import { Link } from 'react-router-dom';
+import { ErrorPageStyled } from './styled';
 
-const NotFoundPage = () => {
+const ErrorPage = () => {
   return (
-    <ErrorMainStyle>
+    <ErrorPageStyled>
       <section>
         <figure>
           <img src={require('../../assets/img/404.webp')} alt={'Error'} />
@@ -18,8 +18,8 @@ const NotFoundPage = () => {
           <Link to="/">BACK TO HOME</Link>
         </figcaption>
       </section>
-    </ErrorMainStyle>
+    </ErrorPageStyled>
   );
 };
 
-export default NotFoundPage;
+export default ErrorPage;
