+++ components/src/pages/Forms/Gender/styled.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const GenderStyled = styled.select`
+  border: 2px solid ${({ theme }) => theme.colors.primary};
+  padding: 5px 12px;
+  border-radius: 100px;
+  font-size: 1.133em;
+
+  &:focus {
+    border-color: ${({ theme }) => theme.colors.secondary};
+    outline: none;
+  }
+`;
