+++ components/src/pages/Forms/SwitchCat/SwitchCat.tsx
@@ -1,5 +1,5 @@
 import React, { forwardRef } from 'react';
-import { SwitchCatStyled } from '../styled';
+import { SwitchCatStyled } from './styled';
 import { SwitchCatProps } from '../Forms.model';
 
 export const SwitchCat = forwardRef<HTMLInputElement, SwitchCatProps>((props, ref) => {
