+++ components/src/pages/Forms/Title/styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const TitleStyled = styled.div`
+  ${({ theme }) => theme.flex.colum}
+  gap: 10px;
+  align-items: center;
+`;
