+++ components/src/pages/Main/Modal/ModalInfo/ModalInfoGenerator.tsx
